# labs/lab03-rest-api/starter/openapi.yaml
openapi: 3.0.3
info:
  title: Lab 3 Item API
  version: 1.0.0
  description: A small HTTP JSON API that holds an array of items that can be added to, deleted from, queried from, or updated.

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /items:
    get:
      summary: List items
      description: Returns the list of items
      responses:
        "200":
          description: Item list response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
              example:
                items:
                  - id: 1
                    name: keyboard
                    quantity: 10
                  - id: 2
                    name: mouse
                    quantity: 5
    post:
      summary: Create a new item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              name: monitor
              quantity: 4
      responses:
        "200":
          description: Item has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              example:
                - id: 3
                  name: monitor
                  quantity: 4
        "400":
          description: Certain data is missing or incorrect.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid or missing data

  /items/{id}:
    get:
      summary: Get item by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID from the URL path.
          schema:
            type: integer
            example: 7
      responses:
        "200":
          description: Single item response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              example:
                - id: 1
                  name: keyboard
                  quantity: 10
        "404":
          description: The item with the provided id was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Item not found
    put:
      summary: Replace the name and quantity of the item with the provided ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID from the URL path.
          schema:
            type: integer
            example: 7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              name: mechanical keyboard
              quantity: 12
      responses:
        "200":
          description: Item has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
              example:
                - id: 1
                  name: mechanical keyboard
                  quantity: 12
        "400":
          description: Certain data is missing or incorrect.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid or missing data
        "404":
          description: The item with the provided id was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Item not found
    delete:
      summary: Delete the item with the provided ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID from the URL path.
          schema:
            type: integer
            example: 7
      responses:
        "204":
          description: Item has been deleted.
        "404":
          description: The item with the provided id was not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Item not found


components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: keyboard
        quantity:
          type: number
          example: 5
      required:
        - id
        - name
        - quantity

    ItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
      required:
        - items

    ItemInput:
      type: object
      properties:
        name:
          type: string
          example: keyboard
        quantity:
          type: number
          example: 5
      required:
        - name
        - quantity

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Item not found
      required:
        - error